Update customize.sh to allow installation on Pixel 10 series
Build ID for 16QPR1 starts with BP3A for all Pixel devices up to Pixel 9 series; Pixel 10 series starts with BD3A. So the assert16QPR function caused installation to fail on Pixel 10 devices. This change simply keeps intact but adds for the grep function to check for either BP3 -OR- BD3, so it shouldn't break anything or open the possibility of PC being installed on an incompatible device.

PS: I don't really use GitHub much and this is I believe my first pull request, so hopefully I'm doing this correctly. Thanks so much for PixelXpert btw, I couldn't live without it!